Convert ImageGallery to TS
- Matches top background colour to page background
