+++ common/model/captioned-image.ts
@@ -1,0 +1,7 @@
+import { ImageType } from './image';
+import { HTMLString } from '../services/prismic/types';
+
+export type CaptionedImage = {
+  caption: HTMLString;
+  image: ImageType;
+};
